🐛 Hotfix

+++ app/src/restaurant/java/de/uni/hannover/hci/mi/team6/covidcheckin/contactForm/ui/ContactFormFragment.kt
@@ -81,7 +81,7 @@ class ContactFormFragment : Fragment() {
                         ServicesModule.beaconMajor,
                         ServicesModule.beaconMajor
                     ),
-                    Address(street, streetNumber.toInt(), zipCode.toInt(), city)
+                    Address(street, streetNumber, zipCode.toInt(), city)
                 )
             )
